+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `MemorError` class
 - `PromptTemplate` class `render` method
+- `PromptTemplate` class `from_content_file` method
 ### Changed
 - `Prompt` class `render` method updated
 - `LLMModel` structure updated